+++ service/src/main/kotlin/fi/espoo/evaka/pis/controllers/PersonController.kt
@@ -10,7 +10,6 @@ import fi.espoo.evaka.identity.isValidSSN
 import fi.espoo.evaka.pis.PersonSummary
 import fi.espoo.evaka.pis.createEmptyPerson
 import fi.espoo.evaka.pis.createPerson
-import fi.espoo.evaka.pis.getDeceasedPeople
 import fi.espoo.evaka.pis.getPersonById
 import fi.espoo.evaka.pis.searchPeople
 import fi.espoo.evaka.pis.service.ContactInfo
@@ -26,13 +25,11 @@ import fi.espoo.evaka.pis.updatePersonContactInfo
 import fi.espoo.evaka.shared.ChildId
 import fi.espoo.evaka.shared.PersonId
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
-import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.BadRequest
 import fi.espoo.evaka.shared.domain.NotFound
 import fi.espoo.evaka.shared.security.AccessControl
 import fi.espoo.evaka.shared.security.Action
-import org.springframework.format.annotation.DateTimeFormat
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PatchMapping
@@ -41,7 +38,6 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import java.time.LocalDate
 
@@ -56,8 +52,7 @@ class PersonController(
     @PostMapping
     fun createEmpty(db: Database, user: AuthenticatedUser): PersonIdentityResponseJSON {
         Audit.PersonCreate.log()
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN, UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.Global.CREATE_PERSON)
         return db.connect { dbc -> dbc.transaction { createEmptyPerson(it) } }
             .let { PersonIdentityResponseJSON.from(it) }
     }
@@ -110,8 +105,7 @@ class PersonController(
         @PathVariable(value = "personId") personId: PersonId
     ): List<PersonWithChildrenDTO> {
         Audit.PersonDependantRead.log(targetId = personId)
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR, UserRole.FINANCE_ADMIN, UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.Person.READ_DEPENDANTS, personId)
         return db.connect { dbc -> dbc.transaction { personService.getPersonWithChildren(it, user, personId) } }?.children
             ?: throw NotFound()
     }
@@ -131,7 +125,7 @@ class PersonController(
     @PostMapping("/search")
     fun findBySearchTerms(
         db: Database,
-        user: AuthenticatedUser,
+        user: AuthenticatedUser.Employee,
         @RequestBody body: SearchPersonBody
     ): List<PersonSummary> {
         Audit.PersonDetailsSearch.log()
@@ -142,7 +136,8 @@ class PersonController(
                     user,
                     body.searchTerm,
                     body.orderBy,
-                    body.sortDirection
+                    body.sortDirection,
+                    restricted = !accessControl.hasPermissionFor(user, Action.Global.SEARCH_PEOPLE_UNRESTRICTED)
                 )
             }
         }
@@ -156,8 +151,7 @@ class PersonController(
         @RequestBody contactInfo: ContactInfo
     ): ContactInfo {
         Audit.PersonContactInfoUpdate.log(targetId = personId)
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR, UserRole.FINANCE_ADMIN, UserRole.STAFF)
+        accessControl.requirePermissionFor(user, Action.Person.UPDATE_CONTACT_INFO, personId)
         return if (db.connect { dbc -> dbc.transaction { it.updatePersonContactInfo(personId, contactInfo) } }) {
             contactInfo
         } else {
@@ -173,8 +167,7 @@ class PersonController(
         @RequestBody data: PersonPatch
     ): PersonJSON {
         Audit.PersonUpdate.log(targetId = personId)
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR, UserRole.FINANCE_ADMIN, UserRole.STAFF)
+        accessControl.requirePermissionFor(user, Action.Person.UPDATE, personId)
 
         val userEditablePersonData = data
             .let {
@@ -203,8 +196,7 @@ class PersonController(
         @PathVariable(value = "personId") personId: PersonId
     ) {
         Audit.PersonDelete.log(targetId = personId)
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.Person.DELETE, personId)
         db.connect { dbc -> dbc.transaction { mergeService.deleteEmptyPerson(it, personId) } }
     }
 
@@ -259,8 +251,7 @@ class PersonController(
         @RequestBody body: GetOrCreatePersonBySsnRequest
     ): PersonJSON {
         Audit.PersonDetailsRead.log()
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR, UserRole.FINANCE_ADMIN)
+        accessControl.requirePermissionFor(user, Action.Global.CREATE_PERSON_FROM_VTJ)
 
         if (!isValidSSN(body.ssn)) throw BadRequest("Invalid SSN")
 
@@ -278,18 +269,6 @@ class PersonController(
             ?: throw NotFound()
     }
 
-    @GetMapping("/get-deceased/")
-    fun getDeceased(
-        db: Database,
-        user: AuthenticatedUser,
-        @RequestParam("sinceDate") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) sinceDate: LocalDate
-    ): List<PersonJSON> {
-        Audit.PersonDetailsRead.log()
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.SERVICE_WORKER, UserRole.UNIT_SUPERVISOR, UserRole.FINANCE_ADMIN)
-        return db.connect { dbc -> dbc.read { it.getDeceasedPeople(sinceDate) } }.map { personDTO -> PersonJSON.from(personDTO) }
-    }
-
     @PostMapping("/merge")
     fun mergePeople(
         db: Database,
@@ -297,8 +276,8 @@ class PersonController(
         @RequestBody body: MergeRequest
     ) {
         Audit.PersonMerge.log(targetId = body.master, objectId = body.duplicate)
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.Person.MERGE, body.master)
+        accessControl.requirePermissionFor(user, Action.Person.MERGE, body.duplicate)
         db.connect { dbc ->
             dbc.transaction { tx ->
                 mergeService.mergePeople(tx, master = body.master, duplicate = body.duplicate)
@@ -313,8 +292,7 @@ class PersonController(
         @RequestBody body: CreatePersonBody
     ): PersonId {
         Audit.PersonCreate.log()
-        @Suppress("DEPRECATION")
-        user.requireOneOfRoles(UserRole.ADMIN, UserRole.SERVICE_WORKER, UserRole.FINANCE_ADMIN)
+        accessControl.requirePermissionFor(user, Action.Global.CREATE_PERSON)
         return db.connect { dbc -> dbc.transaction { createPerson(it, body) } }
     }
 
@@ -326,7 +304,7 @@ class PersonController(
     ) {
         Audit.PersonVtjFamilyUpdate.log(targetId = personId)
         accessControl.requirePermissionFor(user, Action.Person.UPDATE_FROM_VTJ, personId)
-        return db.connect { dbc -> fridgeFamilyService.updatePersonAndFamilyFromVtj(dbc, user.id, personId) }
+        return db.connect { dbc -> fridgeFamilyService.updatePersonAndFamilyFromVtj(dbc, user, personId) }
     }
 
     data class PersonResponse(
